Implement parameter commands and key capture functionality
- Introduced new modules for parameter commands and key capture, enhancing the keybind management experience with customizable command parameters and improved key input handling.
- Updated `app.js` to integrate these new modules, allowing for dynamic parameter input modals and real-time key capture feedback.
- Refactored existing modal management to support the new parameter modal, ensuring a cohesive user experience.
- Added comprehensive tests to validate the new functionality and ensure UI elements reflect changes accurately.

These updates significantly enhance the application's key management capabilities and user interface responsiveness, contributing to a more intuitive user experience.
